+++ lib/utils/utils.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 
 class Clr {
   static Color colorPrimary = Colors.black;
@@ -99,6 +100,8 @@ class Static {
   static Color? btnBorderColor;
   static bool? appDirectionLeftToRight;
   static bool? defaultLoadingProIsIOS;
+  static int stopDecodingFromErrorCode = 400;
+  static Toast toastLength = Toast.LENGTH_SHORT;
 
   Static._();
 }